Move temp storage protection
