+++ shared/js/injected-content-scripts/fingerprint-protection.js
@@ -178,7 +178,6 @@
      */
     function buildInjectionScript () {
         let script = buildScriptProperties()
-        script += modifyTemporaryStorage()
         script += buildBatteryScript()
         script += setWindowDimensions()
         return script
@@ -265,33 +264,6 @@
         }
 
         return windowScript
-    }
-
-    /**
-     * Temporary storage can be used to determine hard disk usage and size.
-     * This will limit the max storage to 4GB without completely disabling the
-     * feature.
-     */
-    function modifyTemporaryStorage () {
-        const script = `
-            if (navigator.webkitTemporaryStorage) {
-                try {
-                    const org = navigator.webkitTemporaryStorage.queryUsageAndQuota
-                    const tStorage = navigator.webkitTemporaryStorage
-                    tStorage.queryUsageAndQuota = function queryUsageAndQuota (callback, err) {
-                        const modifiedCallback = function (usedBytes, grantedBytes) {
-                            const maxBytesGranted = 4 * 1024 * 1024 * 1024
-                            const spoofedGrantedBytes = Math.min(grantedBytes, maxBytesGranted)
-                            callback(usedBytes, spoofedGrantedBytes)
-                        }
-                        org.call(navigator.webkitTemporaryStorage, modifiedCallback, err)
-                    }.bind(null)
-                    Object.defineProperty(Navigator.prototype, 'webkitTemporaryStorage', {get: (() => tStorage).bind(null)})
-                }
-                catch(e) {}
-            }
-        `
-        return script
     }
 
     /**
